Extracted `mdbook test` and `mdbook init` to separate files/modules
