+++ src/bin/mdbook.rs
@@ -6,21 +6,19 @@ extern crate env_logger;
 extern crate open;
 
 use std::env;
-use std::error::Error;
 use std::ffi::OsStr;
 use std::io::{self, Write};
 use std::path::{Path, PathBuf};
 
-use clap::{App, ArgMatches, SubCommand, AppSettings};
+use clap::{App, ArgMatches, AppSettings};
 
 pub mod build;
 pub mod init;
+pub mod test;
 #[cfg(feature = "serve")]
 pub mod serve;
 #[cfg(feature = "watch")]
 pub mod watch;
-
-use mdbook::MDBook;
 
 const NAME: &'static str = "mdbook";
 
@@ -29,16 +27,15 @@ fn main() {
 
     // Create a list of valid arguments and sub-commands
     let app = App::new(NAME)
-                    .about("Create a book in form of a static website from markdown files")
-                    .author("Mathieu David <mathieudavid@mathieudavid.org>")
-                    // Get the version from our Cargo.toml using clap's crate_version!() macro
-                    .version(concat!("v",crate_version!()))
-                    .setting(AppSettings::SubcommandRequired)
-                    .after_help("For more information about a specific command, try `mdbook <command> --help`\nSource code for mdbook available at: https://github.com/azerupi/mdBook")
-                    .subcommand(init::make_subcommand())
-                    .subcommand(build::make_subcommand())
-                    .subcommand(SubCommand::with_name("test")
-                        .about("Test that code samples compile"));
+                .about("Create a book in form of a static website from markdown files")
+                .author("Mathieu David <mathieudavid@mathieudavid.org>")
+                // Get the version from our Cargo.toml using clap's crate_version!() macro
+                .version(concat!("v",crate_version!()))
+                .setting(AppSettings::SubcommandRequired)
+                .after_help("For more information about a specific command, try `mdbook <command> --help`\nSource code for mdbook available at: https://github.com/azerupi/mdBook")
+                .subcommand(init::make_subcommand())
+                .subcommand(build::make_subcommand())
+                .subcommand(test::make_subcommand());
 
     #[cfg(feature = "watch")]
     let app = app.subcommand(watch::make_subcommand());
@@ -53,7 +50,7 @@ fn main() {
         ("watch", Some(sub_matches)) => watch::watch(sub_matches),
         #[cfg(feature = "serve")]
         ("serve", Some(sub_matches)) => serve::serve(sub_matches),
-        ("test", Some(sub_matches)) => test(sub_matches),
+        ("test", Some(sub_matches)) => test::test(sub_matches),
         (_, _) => unreachable!(),
     };
 
@@ -62,29 +59,6 @@ fn main() {
         ::std::process::exit(101);
     }
 }
-
-
-// Simple function that user comfirmation
-fn confirm() -> bool {
-    io::stdout().flush().unwrap();
-    let mut s = String::new();
-    io::stdin().read_line(&mut s).ok();
-    match &*s.trim() {
-        "Y" | "y" | "yes" | "Yes" => true,
-        _ => false,
-    }
-}
-
-
-fn test(args: &ArgMatches) -> Result<(), Box<Error>> {
-    let book_dir = get_book_dir(args);
-    let mut book = MDBook::new(&book_dir).read_config()?;
-
-    book.test()?;
-
-    Ok(())
-}
-
 
 fn get_book_dir(args: &ArgMatches) -> PathBuf {
     if let Some(dir) = args.value_of("dir") {
